requirements.txt
llama-index
llama-index-experimental
fastapi
fastapi[standard]
python-dotenv
pandas